requirements.txt
tensorflow==2.6.0
Keras==2.6.0
numpy==1.19.5
Pillow==8.4.0
firebase-admin==5.0.1
Flask==2.1.0
